Fix cache warmer gs bucket path
